redux store reorganized
