+++ src/features/modal/reducer.js
@@ -1,0 +1,19 @@
+const initialState = {
+  visible: false
+};
+
+const modalReducer = (state = initialState, action) => {
+  switch (action.type) {
+    case "START_COMBAT":
+      return {
+        visible: true,
+        ...action.payload
+      };
+    case "END_COMBAT":
+      return { visible: false };
+    default:
+      return state;
+  }
+};
+
+export default modalReducer;
